third_party/WebKit/LayoutTests/external/wpt/FileAPI/reading-data-section/filereader_abort-expected.txt

This is a testharness.js-based test.
PASS Aborting before read 
FAIL Aborting after read assert_equals: Expected abort event, but got error event instead expected "abort" but got "error"
Harness: the test ran to completion.
